* upstream/developMeertens: (165 commits)
  --Fixed bugs on deleting relation triples
  -- corrected typo in API relation types
  -- fixing formatting ussues
  -- fixing formatting ussues
  -- getting Filters (vor the borwser) is altered so that for tenants their code obtained instead of uri. this is because now filtering on tenats is possibly only via tenant codes, which are indexed in solr. Backwar compatibility should be intact.
  -- restored possibility to filter concepts through setUri (one more possible parameter is added, backward compatibility is intact)
  -- restored field inSkosCollection in solr scheme, without it the filtering on skos collections is not possible
  -- adding parameter to import script: "removeDangling". Removal dangling references happens only if it is explicitely set to "yes".
  -- making get type in Bridge Graph overworking more robust, taking into account that a user can send an object with multiple rdf:type...
  -- updated readme.md
  -- updated apidocs
  -- fixing phpunit.xml
  -- Fixing relationtype code, fixings integration tests for customized settings.
  -- adding relations tests, fixing adding-relation code
  -- fixing autocomplete test and formatting
  -- fixing checking relations for an added concept
  -- trying to fix integration test which works fine "on my machine" but fails on travis
  -- downgrading start-solr.script
  -- fixing initialisation script
  -- grabbing develop version of solrconfig to the developMeertens branch (to be sure) plus downboiling initialisation script to use solr-5.5.4
  ...

+++ application/api/controllers/AbstractController.php
@@ -60,24 +60,40 @@ abstract class AbstractController extends OpenSKOS_Rest_Controller
     {
         $request = $this->getPsrRequest();
         $api = $this->getDI()->make($this->apiResourceClass);
-        $response = $api->create($request);
-        $this->emitResponse($response);
+        $authorisationOn = $this->customAuthorisation($api);
+        if ($authorisationOn || $this->apiResourceClass === 'OpenSkos2\Api\Concept') { 
+            // OpenSkos legacy all API work for concepts withiut auth
+            $response = $api->create($request);
+            $this->emitResponse($response);
+        } else {
+            $this->_501('No authorisation method is specified. POST');
+        }
     }
 
     public function putAction()
     {
         $request = $this->getPsrRequest();
         $api = $this->getDI()->make($this->apiResourceClass);
-        $response = $api->update($request);
-        $this->emitResponse($response);
+        $authorisationOn = $this->customAuthorisation($api);
+        if ($authorisationOn || $this->apiResourceClass === 'OpenSkos2\Api\Concept') {
+            $response = $api->update($request);
+            $this->emitResponse($response);
+        } else {
+            $this->_501('No authorisation method is specified. PUT ');
+        }
     }
 
     public function deleteAction()
     {
         $request = $this->getPsrRequest();
         $api = $this->getDI()->make($this->apiResourceClass);
-        $response = $api->delete($request);
-        $this->emitResponse($response);
+        $authorisationOn = $this->customAuthorisation($api);
+        if ($authorisationOn || $this->apiResourceClass === 'OpenSkos2\Api\Concept') {
+            $response = $api->delete($request);
+            $this->emitResponse($response);
+        } else {
+            $this->_501('No authorisation method is specified. DELETE ');
+        }
     }
 
     /**
@@ -197,6 +213,12 @@ abstract class AbstractController extends OpenSKOS_Rest_Controller
     {
         $parts = RdfNamespace::splitUri($key, false);
         return $parts[1];
+    }
+
+    protected function customAuthorisation($api)
+    {
+        $customInit = $api->getResourceManager()->getCustomInitArray();
+        return isset($customInit['authorisation']);
     }
 
 }
